refactor: 로또 번호 유효성 검증 로직 분리 - Lotto 클래스의 유효성 검증 로직을 validateLottoNumbers 함수로 분리하여 모듈화.
